pyproject.toml
[tool.poetry]
name = "pynalytics-workers"
version = "0.0.1"
description = "Analytics workers with Python"
authors = ["JR <juminrubin@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [
     { include = "shared", from = "src" },
     { include = "speech_to_text_service", from = "src" },
     { include = "summarization_service", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
azure-servicebus = "^7.12.3"
diskcache = "^5.6.3"
openai = "^1.52.0"
whisper = "^1.1.10"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.semantic_release]
branch = "main"
version_variable = "src/__init__.py:__version__"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

